move non-reconcile functions out of machine-controller
